letta/requirements.txt
# requirements.txt
letta-client
chromadb
pypdf
python-dotenv
flask
requests
typing-extensions
google-generativeai

